refactor: remove lokomotive package.
The contents of `cluster.go` mostly deals with talking to the Kubernetes
API to get health status of the Lokomotive cluster and its the only file
in lokomotive package.

Since we already have a package `k8sutil` that deals with talking to the
Kubernetes API, package `k8sutil` seems to be a better place for
`cluster.go`.

This commit moves  the file `pkg/lokomotive/cluster.go` to
`pkg/k8sutil/cluster.go` and updates the package references used
elsewhere in the codebase.
